+++ sdk/models.go
@@ -1,4 +1,4 @@
-package devbridge
+package sdk
 
 // Tunnel 隧道
 type Tunnel struct {